+++ src/main/java/com/rlapcs/radiotransfer/machines/controllers/abstract_controller/AbstractTileController.java
@@ -13,7 +13,7 @@ public abstract class AbstractTileController extends AbstractTileMultiblockNodeW
     public static final int ENCRYPTION_CARD_SLOT_INDEX = 0;
 
     protected static final int ABSTRACT_INVENTORY_SIZE = 1;
-    protected static final double BASE_POWER_USAGE = 10;
+    protected static final int BASE_POWER_USAGE = 10;
 
     protected boolean activated;
     protected int frequency;
@@ -28,7 +28,7 @@ public abstract class AbstractTileController extends AbstractTileMultiblockNodeW
     }
 
     @Override
-    public double getPowerUsagePerTick() {
+    public int getPowerUsagePerTick() {
         return BASE_POWER_USAGE;
     }
